+++ src/components/mainApp/Root.jsx
@@ -3,13 +3,16 @@ import MainApp from "../mainApp/MainApp";
 import UserContextProvider from "../../contexts/UserContext";
 import LocaleContextProvider from "../../contexts/LocaleContext";
 import GlobalContextProvider from "../../contexts/GlobalContext";
+import ClientHydrationContextProvider from "../../contexts/ClientHydrationContext";
 
 function Root() {
   return (
     <GlobalContextProvider>
       <LocaleContextProvider>
         <UserContextProvider>
-          <MainApp />
+          <ClientHydrationContextProvider>
+            <MainApp />
+          </ClientHydrationContextProvider>
         </UserContextProvider>
       </LocaleContextProvider>
     </GlobalContextProvider>